+++ src/cli/index.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-/* global process, __dirname */
+/* global process */
 
 'use strict';
 
@@ -21,34 +21,30 @@ const actions = {
 		action: () => runCommand_preview(true)
 	},
 	preview: {
-		enabled: true,
+		enabled: false,
 		label: `Preview`,
 		description: `Launch Eyas with the current configuration`,
 		command: `preview`,
 		action: runCommand_preview
 	},
-	compile: {
+	bundle: {
 		enabled: true,
-		label: `Compile`,
-		description: `Bundle and compile Eyas with the current configuration for deployment`,
-		command: `compile`,
-		action: runCommand_compile
+		label: `Bundle`,
+		description: `Generate a zipped output for distribution`,
+		command: `bundle`,
+		action: runCommand_bundle
 	}
 };
 
 // setup
 const path = require(`path`);
-const isProd = __dirname.includes(`node_modules`);
+const isDev = process.env.NODE_ENV === `dev`;
 const consumerRoot = process.cwd();
-const moduleRoot = isProd
-	? path.join(consumerRoot, `node_modules`, `@cycosoft`, `eyas`)
-	: consumerRoot;
+const moduleRoot = isDev
+	? consumerRoot
+	: path.join(consumerRoot, `node_modules`, `@cycosoft`, `eyas`);
 const roots = require(path.join(moduleRoot, `dist`, `scripts`, `get-roots.js`));
 const names = {
-	linuxRunner: `linuxRunner.sh`,
-	macRunner: `macRunner.command`,
-	winRunner: `winRunner.cmd`,
-	winDownloader: `getDependency.cmd`,
 	packageJsonCore: `package.json`,
 	packageJson: `package.json`,
 	eyasAssets: `eyas-assets`,
@@ -67,26 +63,42 @@ const paths = {
 	eyasInterfaceDest: path.join(roots.eyasBuild, names.eyasInterface),
 	eyasSrc: path.join(roots.dist, `eyas-core`),
 	eyasDest: roots.eyasBuild,
+	packageJsonModuleSrc: path.join(roots.module, names.packageJson),
 	packageJsonCoreSrc: path.join(roots.dist, `build-assets`, names.packageJsonCore),
 	packageJsonDest: path.join(roots.eyasBuild, names.packageJson),
-	linuxRunnerSrc: path.join(roots.dist, `build-assets`, names.linuxRunner),
-	macRunnerSrc: path.join(roots.dist, `build-assets`, names.macRunner),
-	winRunnerSrc: path.join(roots.dist, `build-assets`, names.winRunner),
-	winRunnerInstallerSrc: path.join(roots.dist, `build-assets`, names.winDownloader),
 	scriptsSrc: path.join(roots.dist, names.scripts),
 	scriptsDest: path.join(roots.eyasBuild, names.scripts),
 	testDest: path.join(roots.eyasBuild, `test`),
-	icon: path.join(roots.eyasBuild, `eyas-assets`, `eyas-logo.png`)
+	icon: path.join(roots.eyasBuild, `eyas-assets`, `eyas-logo.png`),
+	eyasRunnerWinSrc: path.join(roots.dist, `runners`, `Eyas.exe`),
+	eyasRunnerWinDest: path.join(roots.eyasBuild, `Eyas.exe`),
+	macRunnerSrcZip: path.join(roots.dist, `runners`, `Eyas.app.zip`),
+	macRunnerSrc: path.join(roots.dist, `runners`, `Eyas.app`),
+	macRunnerDest: path.join(roots.eyasBuild, `Eyas.app`),
+	linuxRunnerSrc: path.join(roots.dist, `runners`, `Eyas.AppImage`),
+	linuxRunnerDest: path.join(roots.eyasBuild, `Eyas.AppImage`)
 };
 
 // set mode
-actions.previewDev.enabled = !isProd;
+actions.previewDev.enabled = isDev;
 
 // load the user's config
 const config = require(paths.configLoader);
 
 // Entry Point
 (async () => {
+	// ERROR CHECK: capture times when the user's platform isn't supported
+	if (!config.outputs.windows && !config.outputs.mac && !config.outputs.linux) {
+		userWarn(`⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️`);
+		userWarn(`⚠️                                      ⚠️`);
+		userWarn(`⚠️    No supported platforms enabled    ⚠️`);
+		userWarn(`⚠️                                      ⚠️`);
+		userWarn(`⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️  ⚠️`);
+
+		// exit the function
+		return;
+	}
+
 	// import dependencies
 	const { program: cli } = require(`commander`);
 
@@ -136,7 +148,7 @@ function askUser() {
 // setup the CLI arguments
 function defineCommands(cli) {
 	// get the version from the module's package.json
-	const { version } = require(paths.packageJsonCoreSrc);
+	const { version } = require(paths.packageJsonModuleSrc);
 
 	// define the details of the CLI
 	cli
@@ -170,7 +182,6 @@ async function runCommand_config() {
 async function createBuildFolder() {
 	// imports
 	const fs = require(`fs-extra`);
-	const addHours = require(`date-fns/addHours`);
 
 	// give space for the start of the process
 	userLog();
@@ -179,30 +190,49 @@ async function createBuildFolder() {
 	userLog(`Resetting build space...`);
 	await fs.emptyDir(paths.build);
 
-	// copy eyas source to build folder
-	userLog(`Copying Eyas runtime files...`);
-	await fs.copy(paths.eyasSrc, paths.eyasDest);
-	await fs.copy(paths.scriptsSrc, paths.scriptsDest);
-	await fs.copy(paths.eyasInterfaceSrc, paths.eyasInterfaceDest);
-	await fs.copy(paths.eyasAssetsSrc, paths.eyasAssetsDest);
+	// if on Windows, copy the eyas runner to the build folder
+	if(process.platform === `win32`){
+		userLog(`Copying Eyas from ${paths.eyasRunnerWinSrc} to ${paths.eyasRunnerWinDest}...`);
+		await fs.copy(paths.eyasRunnerWinSrc, paths.eyasRunnerWinDest);
+	}
+
+	// if on Mac, copy the eyas runner to the build folder
+	if(process.platform === `darwin`){
+		userLog(`Copying Eyas from ${paths.macRunnerSrc} to ${paths.macRunnerDest}...`);
+		await fs.copy(paths.macRunnerSrc, paths.macRunnerDest);
+	}
+
+	// if on Linux, copy the eyas runner to the build folder
+	if(process.platform === `linux`){
+		userLog(`Copying Eyas from ${paths.linuxRunnerSrc} to ${paths.linuxRunnerDest}...`);
+		await fs.copy(paths.linuxRunnerSrc, paths.linuxRunnerDest);
+	}
 
 	// copy the users source files to the build folder
-	userLog(`Copying test source...`);
+	userLog(`Copying user test from ${path.join(consumerRoot, config.test.source)} to ${paths.testDest}...`);
 	await fs.copy(path.join(consumerRoot, config.test.source), paths.testDest);
 
+	// write the config file
+	const data = getModifiedConfig();
+	await fs.outputFile(paths.configDest, data);
+}
+
+function getModifiedConfig() {
 	// create a new config file with the updated values in the build folder
 	userLog(`Creating snapshot of config...`);
-	delete config.test.source; // isn't used past this point
+	const configCopy = JSON.parse(JSON.stringify(config));
+	delete configCopy.test.source; // isn't used past this point
 
 	// generate meta data for the build
 	const { execSync } = require(`child_process`);
+	const { addHours } = require(`date-fns/addHours`);
 	const now = new Date();
-	const expires = addHours(now, config.outputs.expires);
+	const expires = addHours(now, configCopy.outputs.expires);
 	let gitBranch = ``, gitHash = ``, gitUser = ``;
 	try { gitBranch = execSync(`git rev-parse --abbrev-ref HEAD`).toString().trim(); } catch (e) {/**/}
 	try { gitHash = execSync(`git rev-parse --short HEAD`).toString().trim(); } catch (e) {/**/}
 	try { gitUser = execSync(`git config user.name`).toString().trim(); } catch (e) {/**/}
-	config.meta = {
+	configCopy.meta = {
 		expires,
 		gitBranch,
 		gitHash,
@@ -210,244 +240,114 @@ async function createBuildFolder() {
 		compiled: now
 	};
 
-	// write the config file
-	const data = `module.exports = ${JSON.stringify(config)}`;
-	await fs.outputFile(paths.configDest, data);
+	// wrap the config in a module export
+	const data = `module.exports = ${JSON.stringify(configCopy)}`;
 
 	// let the user know when this build expires
 	userLog(`Set build expirations to: ${expires.toLocaleString()}`);
+
+	// return the updated config data
+	return data;
 }
 
 // launch a preview of the consumers application
 async function runCommand_preview(devMode = false) {
-	const fs = require(`fs-extra`);
 	const { spawn } = require(`child_process`);
-	const electron = require(`electron`);
 
 	// create the build folder to prep for usage
 	await createBuildFolder();
-
-	// copy the package.json to the build folder
-	userLog(`Copying dependency manifest...`);
-	await fs.copy(paths.packageJsonCoreSrc, paths.packageJsonDest);
 
 	// Alert that preview is starting
 	userLog(`Launching preview...`);
 
 	// run the app
-	const command = [paths.eyasApp];
-	if(devMode) { command.push(`--dev`); }
-	spawn(electron, command, {
-		detached: true,
-		stdio: `ignore`,
-		windowsHide: false,
-		cwd: consumerRoot
-	}).unref(); // allow the command line to continue running
+	if(process.platform === `win32`){
+		const command = [];
+		if(devMode) { command.push(`--dev`); }
+		spawn(paths.eyasRunnerWinDest, command, {
+			detached: true,
+			stdio: `ignore`,
+			windowsHide: false,
+			cwd: consumerRoot
+		}).unref(); // allow the command line to continue running
+	} else {
+		const command = [paths.macRunnerDest];
+		// if(devMode) { command.push(`--dev`); }
+		spawn(`open`, command, {
+			detached: true,
+			stdio: `ignore`,
+			windowsHide: false,
+			cwd: consumerRoot
+		}).unref(); // allow the command line to continue running
+
+	}
 
 	// log the end of the process
 	userLog(`Preview launched!`);
 	userLog();
 }
 
-// compile the consumers application for deployment
-async function runCommand_compile() {
-	// imports
+// generate a zipped output for distribution
+async function runCommand_bundle() {
 	const fs = require(`fs-extra`);
-	const builder = require(`electron-builder`);
-	const child_process = require(`child_process`);
+	const archiver = require(`archiver`);
+	const extract = require('extract-zip');
 
-	// create the build folder to prep for usage
-	await createBuildFolder();
-
-	// Clear out the output directory
-	userLog(`Prepare output directory...`);
-	await fs.emptyDir(paths.dist);
-
-	// set the name of the output files
-	const artifactName = `${config.test.title} - ${config.test.version}`;
-
-	// create the portable versions
-	// NOTE: this must come first so installed dependencies for executable aren't included
-	if(config.outputs.portable) {
-		await build_portables();
+	// if the mac runner exists AND it wasn't already extracted
+	if(fs.existsSync(paths.macRunnerSrcZip) && !fs.existsSync(paths.macRunnerSrc)) {
+		// unzip the mac runner
+		await extract(paths.macRunnerSrcZip, { dir: path.join(roots.dist, `runners`) });
 	}
 
-	// create the executable versions
-	if(config.outputs.executable){
-		await build_executables();
-	}
+	// setup the platform output
+	const modifiedConfig = getModifiedConfig();
+	const platforms = [
+		{ enabled: config.outputs.windows, ext: `exe`, runner: paths.eyasRunnerWinSrc, tag: `win` },
+		{ enabled: config.outputs.mac, ext: `app`, runner: paths.macRunnerSrc, tag: `mac` },
+		{ enabled: config.outputs.linux, ext: `AppImage`, runner: paths.linuxRunnerSrc, tag: `linux` }
+	];
 
-	async function build_executables() {
-		// copy the package.json to the build folder
-		userLog();
-		await fs.copy(paths.packageJsonCoreSrc, paths.packageJsonDest);
+	// reset the output directory
+	await fs.emptyDir(roots.eyasDist);
 
-		// Install dependencies
-		userLog(`Installing dependencies...`);
-		child_process.execSync(`npm i`, { cwd: paths.build });
+	// loop through the platforms and create the zipped files if enabled
+	platforms.forEach(platform => {
+		// skip if the platform isn't enabled
+		if(!platform.enabled) { return; }
 
-		userLog(`Creating "executable" distributable(s)...`);
-		userLog();
+		const artifactName = `${config.test.title} - ${config.test.version}.${platform.tag}.zip`;
 
-		// Build the executables
-		const targets = [];
-		if(config.outputs.windows) { targets.push(builder.Platform.WINDOWS); }
-		if(config.outputs.mac) { targets.push(builder.Platform.MAC); }
-		if(config.outputs.linux) { targets.push(builder.Platform.LINUX); }
+		// create the zip file
+		const output = fs.createWriteStream(path.join(roots.eyasDist, artifactName));
 
-		const builtFiles = await builder.build({
-			targets: targets.length ? builder.createTargets(targets) : null,
-			config: {
-				appId: `com.cycosoft.eyas`,
-				productName: `Eyas`,
-				// eslint-disable-next-line quotes
-				artifactName: `${artifactName}` + '.${ext}',
-				copyright: `Copyright © 2023 Cycosoft, LLC`,
-				asarUnpack: [`resources/**`],
-				directories: {
-					app: paths.build,
-					output: paths.dist
-				},
-				removePackageScripts: true,
-				removePackageKeywords: true,
-				mac: {
-					target: `dmg`,
-					icon: paths.icon
-					// identity: `undefined` // disable code signing
-				},
-				win: {
-					target: `portable`,
-					icon: paths.icon
-				},
-				linux: {
-					target: `AppImage`,
-					icon: paths.icon,
-					category: `Utility`
-				}
-			}
+		// when the process has completed (Note: listener must be added before the archive is finalized)
+		output.on(`close`, () => {
+			// alert the user
+			userLog(`🎉 File created -> ${artifactName}`);
 		});
 
-		// wrap the executables in a zip file
-		const archiver = require(`archiver`);
-		let completedZipCount = 0;
-		builtFiles.forEach(file => {
-			// skip if the file is a blockmap
-			if(file.endsWith(`.blockmap`)) {
-				completedZipCount++;
-				return;
-			}
+		// prepare a new archive
+		const archive = archiver(`zip`, { store: isDev, zlib: { level: 9 } });
 
-			// if the file is .AppImage, do not archive but let the user know it was created
-			if(file.endsWith(`.AppImage`)) {
-				completedZipCount++;
-				userLog(`🎉 File created -> ${file}`);
-				return;
-			}
+		// push content to the archive
+		archive.pipe(output);
 
-			// create the zip file
-			const output = fs.createWriteStream(`${file}.zip`);
-			output.on(`close`, () => {
-				completedZipCount++;
-				userLog(`🎉 File created -> ${file}.zip`);
-
-				if(completedZipCount === builtFiles.length) {
-					performCleanup();
-				}
-			});
-			const archive = archiver(`zip`, { store: true });
-			archive.pipe(output);
-			const filename = file.split(`\\`).pop();
-			archive.file(file, { name: filename });
-			archive.finalize();
-		});
-
-		async function performCleanup() {
-			// delete directories in the build output
-			// delete files that aren't .zip, .AppImage
-			userLog(`Performing cleanup...`);
-			const files = await fs.readdir(paths.dist);
-			for(const file of files) {
-				// skip file if it's in the skip list
-				let shouldSkip = false;
-				const skipList = [`.zip`, `.AppImage`];
-				skipList.forEach(skip => {
-					if(file.endsWith(skip)) { shouldSkip = true; }
-				});
-
-				// exit this loop if the file should be skipped
-				if(shouldSkip) { continue; }
-
-				// get the full path to the file
-				const filePath = path.join(paths.dist, file);
-
-				// if it's a directory, delete it
-				if((await fs.stat(filePath)).isDirectory()) {
-					await fs.remove(filePath);
-					continue;
-				}
-
-				// delete the file
-				await fs.remove(filePath);
-			}
-
-			// log the end of the process
-			userLog(`Executable compilation complete!`);
-			userLog();
+		// add the appropriate runner
+		if(platform.tag === `mac`) {
+			archive.directory(platform.runner, `Eyas.${platform.ext}`);
+		} else {
+			archive.file(platform.runner, { name: `Eyas.${platform.ext}` });
 		}
-	}
 
-	async function build_portables() {
-		// overwrite the installer manifest with the runner manifest
-		userLog();
-		userLog(`Copying dependency manifest...`);
-		await fs.copy(paths.packageJsonCoreSrc, paths.packageJsonDest);
+		// add the updated config
+		archive.append(modifiedConfig, { name: `.eyas.config.js` });
 
-		// for each OS, create the runner file
-		[`mac`, `windows`, `linux`].forEach(os => {
-			// exit if the OS isn't enabled
-			if(!config.outputs[os]) { return; }
+		// add the user's test
+		archive.directory(path.join(consumerRoot, config.test.source), `test`);
 
-			// create the node runner version
-			userLog(`Creating ${os.toUpperCase()} "portable" distributable...`);
-			const filename = `${artifactName}.${os}.zip`;
-			const archiver = require(`archiver`);
-			const output = fs.createWriteStream(paths.dist + `/${filename}`);
-			const archive = archiver(`zip`, { zlib: 9 });
-			output.on(`close`, () => {
-				userLog(`🎉 File created -> ${path.join(paths.dist, filename)}`);
-			});
-			archive.pipe(output);
-
-			// add common files
-			archive.directory(paths.build, `app`);
-
-			// name of the file that runs Eyas
-			const runnerName = `Run Test`;
-
-			// add linux files
-			if (os === `linux`) {
-				const linuxFileExt = names.linuxRunner.split(`.`).pop();
-				archive.file(paths.linuxRunnerSrc, { name: `${runnerName}.${linuxFileExt}` });
-			}
-
-			// add mac files
-			if (os === `mac`) {
-				const macFileExt = names.macRunner.split(`.`).pop();
-				archive.file(paths.macRunnerSrc, { name: `${runnerName}.${macFileExt}` });
-			}
-
-			// add win files
-			if (os === `windows`) {
-				const winFileExt = names.winRunner.split(`.`).pop();
-				archive.file(paths.winRunnerSrc, { name: `${runnerName}.${winFileExt}` });
-				archive.file(paths.winRunnerInstallerSrc, { name: names.winDownloader });
-			}
-
-			// complete the archive
-			archive.finalize();
-		});
-
-	}
+		// close the archive
+		archive.finalize();
+	});
 }
 
 // wrapper to differentiate user logs (allowed) from system logs (disallowed)
@@ -457,4 +357,10 @@ function userLog(string) {
 
 	// eslint-disable-next-line no-console
 	console.log(output);
+}
+
+// wrapper to avoid linting errors
+function userWarn(input) {
+	// eslint-disable-next-line no-console
+	console.warn(input);
 }