+++ src/scripts/compile-module.js
@@ -1,5 +1,7 @@
 #!/usr/bin/env node
 
+/* global process */
+
 'use strict';
 
 // Import dependencies
@@ -12,79 +14,71 @@ const names = {
 	buildAssets: `build-assets`,
 	eyasAssets: `eyas-assets`,
 	eyasInterfaceApp: `eyas-interface`,
+	packageJson: `package.json`,
 	cli: `cli`
 };
 const paths = {
 	dist: roots.dist,
-	buildAssetsSrc: path.join(roots.src, names.buildAssets),
-	buildAssetsDest: path.join(roots.dist, names.buildAssets),
+	buildAssetsSrc: path.join(roots.src, names.buildAssets, names.packageJson),
+	buildAssetsDest: path.join(roots.moduleBuild, names.packageJson),
 	cliDest: path.join(roots.dist, names.cli),
 	cliSrcFile: path.join(roots.src, names.cli, `index.js`),
 	cliDestFile: path.join(roots.dist, names.cli, `index.js`),
+	eyasCoreSrc: path.join(roots.preBuild, `eyas-core/index.js`),
+	eyasCoreDest: path.join(roots.moduleBuild, `index.js`),
 	eyasAssetsSrc: path.join(roots.src, names.eyasAssets),
-	eyasAssetsDest: path.join(roots.dist, names.eyasAssets),
-	eyasInterfaceAppSrc: path.join(roots.src, names.eyasInterfaceApp, `app`, `dist`),
-	eyasInterfaceAppDest: path.join(roots.dist, names.eyasInterfaceApp),
-	packageJsonModule: path.join(roots.module, `package.json`),
-	packageJsonDist: path.join(roots.dist, names.buildAssets, `package.json`)
+	eyasAssetsDest: path.join(roots.moduleBuild, names.eyasAssets),
+	eyasInterfaceAppSrc: path.join(roots.preBuild, names.eyasInterfaceApp),
+	eyasInterfaceAppDest: path.join(roots.moduleBuild, names.eyasInterfaceApp),
+	packageJsonModule: path.join(roots.module, names.packageJson),
+	packageJsonDist: path.join(roots.moduleBuild, names.packageJson),
+	scriptsSrc: path.join(roots.preBuild, `scripts`),
+	scriptsBuild: path.join(roots.moduleBuild, `scripts`),
+	scriptsDist: path.join(roots.dist, `scripts`)
 };
 
 // Allow for "root" await calls
 (async () => {
-	// Prep the dist/ directory for module output
-	await fs.emptyDir(paths.dist);
+	// Prep the .build/ & dist/ directories for module output
+	// await fs.emptyDir(roots.moduleBuild);
+
+	// if(process.env.FORCE_BUILD !== `win32`) {
+	// 	await fs.emptyDir(roots.dist);
+	// }
 
 	// Copy runtime files
 	await fs.copy(paths.eyasAssetsSrc, paths.eyasAssetsDest);
 	await fs.copy(paths.buildAssetsSrc, paths.buildAssetsDest);
 	await fs.copy(paths.eyasInterfaceAppSrc, paths.eyasInterfaceAppDest);
+	await fs.copy(paths.eyasCoreSrc, paths.eyasCoreDest);
+	await fs.copy(paths.scriptsSrc, paths.scriptsBuild);
 
-	// set the npm and node dependency version numbers based on package.json
-	await updateRunnerVersions();
+	await fs.copy(
+		path.join(paths.scriptsSrc, `get-config.js`),
+		path.join(paths.scriptsDist, `get-config.js`)
+	);
+
+	await fs.copy(
+		path.join(paths.scriptsSrc, `get-roots.js`),
+		path.join(paths.scriptsDist, `get-roots.js`)
+	);
+
+	// remove the prebuild folder
+	// console.log(`clean up: removing ${roots.preBuild}`);
+	// await fs.remove(roots.preBuild);
 
 	// Update the package.json version numbers
 	await updatePackageJsonValues();
 
-	// Compile the CLI
-	console.log(`🕜 compile-module.js emptying ${paths.cliDest}`);
-	await fs.emptyDir(paths.cliDest);
-	console.log(`🕜 compile-module.js compiling ${paths.cliSrcFile}`);
+	// Copy the CLI
+	console.log(`copying ${paths.cliSrcFile}`);
 	await fs.copy(paths.cliSrcFile, paths.cliDestFile);
-	console.log(`🎉 compile-module.js complete`);
+	console.log(`🎉 complete`);
 })();
-
-// set the npm and node dependency version numbers for each runner
-async function updateRunnerVersions() {
-	console.log(`🕜 updateRunnerVersions() start`);
-	// read the package.json
-	const packageJson = require(paths.packageJsonModule);
-	const nodeVersion = packageJson.devDependencies.node.match(/\d+\.\d+\.\d+/)[0];
-	const npmVersion = packageJson.devDependencies.npm.match(/\d+\.\d+\.\d+/)[0];
-	const runners = [`winRunner.cmd`, `macRunner.command`, `linuxRunner.sh`];
-
-	// for each runner
-	for (const filename of runners) {
-		const runnerPath = path.join(paths.buildAssetsDest, filename);
-
-		// read the contents of the runner
-		let script = fs.readFileSync(runnerPath, `utf8`);
-
-		// modify the nodeVersion variable
-		script = script
-			.replace(/nodeVersion=0.0.0/, `nodeVersion=${nodeVersion}`)
-			.replace(/npmVersion=0.0.0/, `npmVersion=${npmVersion}`);
-
-		// save the modified runner
-		console.log(`📝 updateRunnerVersions() writing ${filename}`);
-		await fs.outputFile(runnerPath, script);
-	}
-
-	console.log(`🕜 updateRunnerVersions() end`);
-}
 
 // update all the versions in the distributed package.json from the module package.json
 async function updatePackageJsonValues() {
-	console.log(`🕜 updatePackageJsonValues() start`);
+	console.log(`🕜 updatePackageJsonValues()`);
 	// read both package.json
 	const packageJsonModule = require(paths.packageJsonModule);
 	const packageJsonDist = require(paths.packageJsonDist);
@@ -120,8 +114,5 @@ async function updatePackageJsonValues() {
 	packageJsonDist.devDependencies.electron = packageJsonModule.dependencies.electron;
 
 	// save the updated dist/package.json
-	console.log(`📝 updatePackageJsonValues() writing package.json`);
 	await fs.outputFile(paths.packageJsonDist, JSON.stringify(packageJsonDist));
-
-	console.log(`🕜 updatePackageJsonValues() end`);
 }